refactor: used Organism Alias Path
