+++ client/src/Component/Organism/Chat/ChatListContainer.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { css } from "@emotion/react";
-import ChatImageContainer from "./ChatImageContainer";
-import ChatListDetail from "./ChatListDetail";
+import { ChatImageContainer } from "./ChatImageContainer";
+import { ChatListDetail } from "./ChatListDetail";
 import { ChatListContainerType } from "../../../Util/type";
 import { handleModalClick } from "../../../Util";
 
@@ -14,13 +14,11 @@ const ChatListContainerStyle = css`
   min-width: 520px;
 `;
 
-function ChatListContainer({ profileRef, setOpenModal }: ChatListContainerType) {
+export const ChatListContainer = ({ profileRef, setOpenModal }: ChatListContainerType) => {
   return (
     <div css={ChatListContainerStyle} onClick={(e) => handleModalClick(e, profileRef, setOpenModal)}>
       <ChatImageContainer profileRef={profileRef} />
       <ChatListDetail />
     </div>
   );
-}
-
-export default ChatListContainer;
+};
