+++ client/src/Component/Organism/Chat/ChatProfile.tsx
@@ -6,7 +6,8 @@ import { ChatProfileInfoContainer } from "@Molecules/Chat/ChatProfileInfoContain
 import { ChatInfoType } from "@Util/type";
 const dummyImage = "Asset/meetingImage.png";
 
-export default function ChatProfile({ data, idx, chatRoomRef }: { data: ChatInfoType; idx: number; chatRoomRef: React.MutableRefObject<HTMLDivElement[]> }) {
+type props = { data: ChatInfoType; idx: number; chatRoomRef: React.MutableRefObject<HTMLDivElement[]> };
+export const ChatProfile = ({ data, idx, chatRoomRef }: props) => {
   const memberType = data.member.length > 2 ? "team" : data.member[0].sex;
   const lastChatInfo = data.chatMessage[data.chatMessage.length - 1];
   return (
@@ -19,4 +20,4 @@ export default function ChatProfile({ data, idx, chatRoomRef }: { data: ChatInfo
       </ProfileCard>
     </div>
   );
-}
+};
