+++ client/src/Component/Organism/Info/TeamInfo.tsx
@@ -1,6 +1,6 @@
 import React, { RefObject } from "react";
 import { css } from "@emotion/react";
-import TeamInfoContainer from "./TeamInfoContianer";
+import { TeamInfoContainer } from "./TeamInfoContianer";
 import { InfoImageContainer } from "@Hoc/.";
 
 const TeamInfoContainerStyle = css`
@@ -13,11 +13,12 @@ const TeamInfoContainerStyle = css`
   align-items: center;
 `;
 
-export default function TeamInfo({ teamNameRef, teamInfoRef, setLocSelected }: { teamNameRef: RefObject<HTMLInputElement>; teamInfoRef: RefObject<HTMLInputElement>; setLocSelected: (value: string) => void }) {
+type props = { teamNameRef: RefObject<HTMLInputElement>; teamInfoRef: RefObject<HTMLInputElement>; setLocSelected: (value: string) => void };
+export const TeamInfo = ({ teamNameRef, teamInfoRef, setLocSelected }: props) => {
   return (
     <div css={TeamInfoContainerStyle}>
       <InfoImageContainer />
       <TeamInfoContainer setLocSelected={setLocSelected} teamNameRef={teamNameRef} teamInfoRef={teamInfoRef} />
     </div>
   );
-}
+};
