+++ client/src/Component/Organism/index.tsx
@@ -2,3 +2,14 @@ export * from "./MainPage/MainHeader";
 export * from "./MainPage/MainBody";
 export * from "./MainPage/MainBodyLeft";
 export * from "./MainPage/MainBodyRight";
+
+export * from "./Chat/ChatDetail";
+export * from "./Chat/ChatImageContainer";
+export * from "./Chat/ChatInputContainer";
+export * from "./Chat/ChatListContainer";
+export * from "./Chat/ChatListDetail";
+export * from "./Chat/ChatProfile";
+export * from "./Chat/ChatProfileContainer";
+
+export * from "./Info/MyInfoContainer";
+export * from "./Info/TeamInfo";
